external/SABLE/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
python-chess>=1.9.0
scikit-learn>=1.0.0
tqdm>=4.62.0
rich>=12.0.0
psutil>=5.8.0 